+++ kapusha/render/Shader.cpp
@@ -7,7 +7,7 @@ shader_t::shader_t(const core::string_desc_t &string, type_e type)
   : type_(type) {
   name_ = glCreateShader(static_cast<GLenum>(type_));
   const char *str = string.begin();
-  const GLint len = string.length();
+  const GLint len = static_cast<int>(string.length());
   glShaderSource(name_, 1, &str, &len);
   glCompileShader(name_);
   glGetShaderiv(name_, GL_COMPILE_STATUS, &compiled_);
